+++ lib/amwal_api.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'amwal_api_platform_interface.dart';
 import 'package:flutter/services.dart';
 
@@ -5,7 +7,7 @@ class AmwalApi {
   final String merchantId;
   final String countryCode;
   final String phoneNumber;
-  final double amount;
+  final Float amount;
   final MethodChannel _channel;
 
   AmwalApi._({
@@ -19,11 +21,13 @@ class AmwalApi {
   factory AmwalApi({required AmwalApiBuilder builder}) {
     final channel = MethodChannel('amwal_api');
     return builder
-        .startPayment(builder._merchantId, builder._countryCode, builder._phoneNumber, builder._amount)
+        .startPayment(builder._merchantId, builder._countryCode,
+            builder._phoneNumber, builder._amount)
         .build(channel: channel);
   }
-  
-  Future<String> startPayment(String merchantId, String countryCode, String phoneNumber, double amount) async {
+
+  Future<String> startPayment(String merchantId, String countryCode,
+      String phoneNumber, Float amount) async {
     final Map<String, dynamic> args = <String, dynamic>{
       'merchantId': merchantId,
       'countryCode': countryCode,
@@ -32,13 +36,12 @@ class AmwalApi {
     };
     final dynamic result = await _channel.invokeMethod('startPayment', args);
     return result as String;
-
   }
 
   Future<String?> getPlatformVersion() {
     return AmwalApiPlatform.instance.getPlatformVersion();
   }
-  
+
   Map<String, String> _adapter() {
     return {
       'merchantId': merchantId,
@@ -49,18 +52,19 @@ class AmwalApi {
   }
 }
 
-
 class AmwalApiBuilder {
   String _merchantId = "";
   String _countryCode = "";
   String _phoneNumber = "";
-  double _amount = 0.0;
+  //add float value in amount
+  Float _amount = 0.0 as Float;
 
-  AmwalApiBuilder startPayment(String merchantId, String countryCode, String phoneNumber, double amount){
-    _merchantId=  merchantId;
-    _countryCode= countryCode;
-    _phoneNumber= phoneNumber;
-    _amount= amount; 
+  AmwalApiBuilder startPayment(
+      String merchantId, String countryCode, String phoneNumber, Float amount) {
+    _merchantId = merchantId;
+    _countryCode = countryCode;
+    _phoneNumber = phoneNumber;
+    _amount = amount;
     return this;
   }
 
@@ -68,7 +72,7 @@ class AmwalApiBuilder {
     /* if (_merchantId.isEmpty) {
      throw Exception("Merchant Id has to be provided");
     }*/
-    
+
     return AmwalApi._(
       merchantId: _merchantId,
       countryCode: _countryCode,
